requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Pillow==10.0.0
stripe==7.8.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
werkzeug==3.0.1
gunicorn==21.2.0 